renamed NavBar to TopBar to avoid conflicts with Bootstrap's Navbar
