+++ src/components/Settings.js
@@ -1,0 +1,7 @@
+import TopBar from "./TopBar";
+
+function Settings() {
+  return <></>;
+}
+
+export default Settings;
